chore: refactor ResponsiveProvider
