+++ .eslintrc.json
@@ -10,6 +10,7 @@
         "eslint:recommended",
         "plugin:prettier/recommended",
         "plugin:react/recommended",
+        "plugin:react-hooks/recommended",
         "prettier/react"
     ],
     "parser": "babel-eslint",
