+++ package.json
@@ -70,6 +70,7 @@
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^4.0.8",
     "eslint-plugin-sort-imports-es6-autofix": "^0.5.0",
     "husky": "^2.4.0",
     "jest": "^24.8.0",
